#30 - [refact] Controller 리팩토링
- Rest API와 API의 Controller 분리
  - Rest API Contoller의 경우, url 앞에 "/api"를 추가하였고, Controller에 Api라는 이름을 추가함
  - View를 반환하는 Controller의 경우, url은 그대로 유지하되, Controller에 Page라는 이름을 추가함
